refactor: split cli.py into commands/ package (#127)
Move all 18+ CLI commands from monolithic cli.py (5719 lines) into
7 focused modules under src/klemma/commands/. Shared helpers, the main
Click group, and the console instance remain in cli.py (1609 lines).

Commands import from ..cli (no circular imports — cli.py imports
commands at the bottom). All 918 tests pass, ruff clean.

Module breakdown:
- acquire.py (169 lines): acquire command
- analyze.py (487 lines): status, suggest, gaps, source groups
- benchmark.py (502 lines): benchmark command
- manage.py (1235 lines): init, plan, info, tree, reassign, migrate
- process.py (577 lines): process, embed group, similar
- research.py (771 lines): research, ask, library group
- write.py (516 lines): draft group, outline
