+++ src/klemma/commands/__init__.py
@@ -1,0 +1,7 @@
+"""Command modules for the Klemma CLI.
+
+Each module registers its commands/groups against ``main`` from ``..cli``.
+Importing this package triggers registration of all commands.
+"""
+
+from . import acquire, analyze, benchmark, manage, process, research, write  # noqa: F401
